Bug 1035892 - Handle 64-bit addresses for EXC_BAD_ACCESS exceptions on Mac r=froydnj
Differential Revision: https://phabricator.services.mozilla.com/D24356

UltraBlame original commit: 2b1ffc8004d5e8160bc5bb1109c9f67a003048ef

+++ toolkit/crashreporter/breakpad-client/mac/handler/exception_handler.h
@@ -104,7 +104,7 @@ class ExceptionHandler {
   typedef bool (*DirectCallback)( void *context,
                                   int exception_type,
                                   int exception_code,
-                                  int exception_subcode,
+                                  int64_t exception_subcode,
                                   mach_port_t thread_name);
 
   
@@ -203,7 +203,7 @@ class ExceptionHandler {
   
   bool WriteMinidumpWithException(int exception_type,
                                   int exception_code,
-                                  int exception_subcode,
+                                  int64_t exception_subcode,
                                   breakpad_ucontext_t *task_context,
                                   mach_port_t thread_name,
                                   bool exit_after_write,
